DBI methods

Implementations of pure virtual functions defined in the `DBI` package.

# S4 method for OsqueryDriver
show(object)

# S4 method for OsqueryDriver
dbConnect(drv, host = NULL, keyfile = NULL,
  osquery_remote_path = NULL, ...)

# S4 method for OsqueryDriver,ANY
dbDataType(dbObj, obj, ...)

# S4 method for OsqueryDriver,list
dbDataType(dbObj, obj, ...)

# S4 method for OsqueryDriver
dbIsValid(dbObj, ...)

# S4 method for OsqueryConnection
show(object)

# S4 method for OsqueryConnection
dbIsValid(dbObj, ...)

# S4 method for OsqueryConnection
dbDisconnect(conn, ...)

# S4 method for OsqueryConnection,character
dbSendQuery(conn, statement, ...)

# S4 method for OsqueryConnection,character
dbSendStatement(conn, statement, ...)

# S4 method for OsqueryConnection,ANY
dbDataType(dbObj, obj, ...)

# S4 method for OsqueryConnection,character
dbQuoteString(conn, x, ...)

# S4 method for OsqueryConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for OsqueryConnection,character
dbReadTable(conn, name, ...)

# S4 method for OsqueryConnection
dbListTables(conn, ...)

# S4 method for OsqueryConnection,character
dbExistsTable(conn, name, ...)

# S4 method for OsqueryConnection,character
dbListFields(conn, name, ...)

# S4 method for OsqueryResult
show(object)

# S4 method for OsqueryResult
dbClearResult(res, ...)

# S4 method for OsqueryResult
dbFetch(res, n = -1, ...)

# S4 method for OsqueryResult
dbHasCompleted(res, ...)

# S4 method for OsqueryResult
dbIsValid(dbObj, ...)

# S4 method for OsqueryResult
dbGetStatement(res, ...)

Arguments

object
Any R object

drv
an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

host
NULL (default) if using osqueryi locally; an ssh server string of the form [user@]hostname[:@port]

keyfile
NULL (default) if using osqueryi locally; path to private key file. Must be in OpenSSH format (see details)

osquery_remote_path
if connecting to a remote system and osqueryi is not on the PATH for the ssh connection, provide it here. Ignored if using locally.

...
authentication arguments needed by the DBMS instance; these typically include user, password, host, port, dbname, etc. For details see the appropriate DBIDriver.

dbObj
A object inheriting from DBIDriver or DBIConnection

obj
An R object whose SQL type we want to determine.

conn
A DBIConnection object, as returned by dbConnect().

statement
a character string containing SQL.

x
A character vector to quote as string.

name
A character string specifying the unquoted DBMS table name, or the result of a call to dbQuoteIdentifier().

res
An object inheriting from DBIResult.

n
maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.